+++ helper/widgets/markdown.js
@@ -28,13 +28,11 @@ Markdown.constitute(function prepareSchema() {
  *
  * @author   Jelle De Loecker   <jelle@elevenways.be>
  * @since    0.1.0
- * @version  0.1.0
+ * @version  0.2.1
  *
  * @param    {HTMLElement}   widget
  */
 Markdown.setMethod(function populateWidget() {
-
-	populateWidget.super.call(this);
 
 	let source = this.config.markdown || '';