+++ helper/widgets/table_of_contents.js
@@ -16,11 +16,9 @@ const Toc = Function.inherits('Alchemy.Widget', 'TableOfContents');
  *
  * @author   Jelle De Loecker   <jelle@elevenways.be>
  * @since    0.1.2
- * @version  0.2.0
+ * @version  0.2.1
  */
 Toc.setMethod(function populateWidget() {
-
-	populateWidget.super.call(this);
 
 	let toc = this.createElement('al-toc');